Avoid S.S.Permissions dependency in S.S.Crypto.Xml
Avoid the S.S.Permissions dependency by referencing the
latest available System.Security.AccessControl package
which contans the Evidence types in all tfms in the
System.Security.Cryptography.Xml project.

Add suppression file for Permissions
